plasmoidviewer: respect --containment flag in sizing and scenerect
The normal behavior of plasmoidviewer, which resizes the applet to the
window, is not very useful for testing containments, at it sets the
scenerect and resizes the widget where it is really the containment's
job to take care of the layout.

So, when the --containment flag is set, and a plugin is found:
- do not change the scenerect anymore
- do not resize the widget
- do not wire up applet resize events to our view's scenerect
- set window title to $Applet: $Containment
- change icon to containment's icon

This set of changes makes plasmoidviewer really useful, where it
previously would often flick out of the view and show all kinds of
applet placement madness when used with the --containment flag.

REVIEW:108284