Make konsolepreview to work even if the plasmatepreviewerlog.txt doesn't exist.
KStandarDirs::locate returns a path only if the file exists, the first time that
we run plasmate the plasmatepreviewerlog.txt doesn't exist so the konsolepreviewer doesn't work.
In order to fix this we are replacing the KStandarDirs:locate with KStandarDirs::locateLocal.
KStandarDirs::locateLocal always returns a path. Even if the file doesn't exist.
REVIEW: 108679