Checks if an optional field was instantiated before accessing it
The private field m_konsolewidget is initialized only on some specific cases, so it should be accessed with caution. But there are some places in witch it's being used without checking if it was instantiated or not. This patch fixes those accesses, by checking if it is instantiated or not.

This was laying around in reviewboard approved by Giorgos Tsiapaliokas but uncommited for 5 months

REVIEW: 110288
BUGS: 319291