pointers -> QWeakPointer
we can't be sure about the lifetime of the widgets since the signal
loadRequiredEditor will pass them to QMainWindow::setCentralWidget
and the QMainWindow will delete them in the appropriate time. So
we will use QWeakPointer in order to be sure about the state of the
widgets.
The deletion of the widgets from the QMainWindow is a desired behaviour
since our user may use x of those widgets just *once* and then return to the
texteditor so there is no reason to waste extra memory