[Cuttlefish] Add 48 to the sizeslider
Summary:
The initial icon size Cuttlefish starts with is 48 (Kirigami.Unis.iconSizes.large).
However this size cannot be selected using the slider. To quickly see this
start Cuttlefish (note the slider says 64) and move the slide to some other
value and back to 64. The icons have now a different size than before. Since we
also show a preview for 48 add this size to the selectable ones.

Reviewers: #plasma, cblack, ngraham

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D24304