[Cuttlefish] Use the same actions in all menus
Summary:
Defines the actions in one place and uses them for the context menus and
buttons in the two preview panes. This reduces code duplication, adds
actions to the context menu which couldn't be triggered from there before
and makes it easier to add new features with only needing to touch one place
instead of potentially three. Furthermore this also reduces the risk of bugs
caused by copy-pasting, for example the screenshot actions didn't work before
in responsive mode.
The actionsMenu component is based on the one from Kirigami with the same name.

Test Plan: Everything works

Reviewers: #plasma, ngraham, cblack

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25846