engineexplorer: Unify first two columns (DataSource + Key)
They looked weird and broken with those empty cells on the left. In
every possible scenario only one of them is filled at a time. Also,
without moving keys to the left indented column, hierarchical data looks
especially plain and wrong.

Width of the first column is also manually increased to be twice as wide
as the stock value of 100, to accommodate long keys and deep nesting.