plasmoidviewer: Remove remnants of "konsole"
If my observation skills are good, this feature never worked for at
least almost a decade. Maybe it worked in a different way before the
QML port, I don't really know. But as of today, it surely does nothing,
reacts to nothing, displays nothing (even if it could), and is hardcoded
to be disabled.