kqml: Add hot reload support
When working on testing QML snippets, having to save and close and
reopen is a drag.
This patch adds support to reload any QML when the file on disk changes,
making it suitable
for rapid prototyping and testing

As it's a relatively new dev tool, this is on by default. We can make it
an option if it proves
annoying.  A behavioural change this introduces is that invalid QML will
open a window
instead of quitting.